+++ packages/designer/src/index.ts
@@ -4,3 +4,4 @@ export * from './designer';
 export * from './document';
 export * from './project';
 export * from './builtin-simulator';
+export * from './plugin';