+++ packages/editor-skeleton/src/transducers/parse-props.ts
@@ -69,7 +69,7 @@ function propTypeToSetter(propType: PropType): SetterType {
       const componentName = dataSource.length >= 4 ? 'SelectSetter' : 'RadioGroupSetter';
       return {
         componentName,
-        props: { dataSource },
+        props: { dataSource, options: dataSource },
         isRequired,
         initialValue: dataSource[0] ? dataSource[0].value : null,
       };
@@ -139,6 +139,11 @@ function propTypeToSetter(propType: PropType): SetterType {
         componentName: 'FunctionSetter',
         isRequired,
       };
+    case 'color':
+      return {
+        componentName: 'ColorSetter',
+        isRequired,
+      };
     case 'oneOfType':
       return {
         componentName: 'MixedSetter',
@@ -148,6 +153,8 @@ function propTypeToSetter(propType: PropType): SetterType {
         },
         isRequired,
       };
+    default:
+      // do nothing
   }
   return {
     componentName: 'MixedSetter',