+++ packages/editor-core/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.27"></a>
+## [1.0.27](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/v1.0.27-beta.2...v1.0.27) (2020-12-24)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-editor-core
+
+<a name="1.0.27-beta.2"></a>
+## [1.0.27-beta.2](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/v1.0.27-beta.1...v1.0.27-beta.2) (2020-12-23)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-editor-core
+
+<a name="1.0.27-beta.1"></a>
+## [1.0.27-beta.1](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/v1.0.27-beta.0...v1.0.27-beta.1) (2020-12-23)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-editor-core
+
+<a name="1.0.27-beta.0"></a>
+## [1.0.27-beta.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/v1.0.26...v1.0.27-beta.0) (2020-12-23)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-editor-core
+
+<a name="1.0.26"></a>
+## [1.0.26](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/v1.0.26-beta.1...v1.0.26) (2020-12-22)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-editor-core
+
+<a name="1.0.26-beta.1"></a>
+## [1.0.26-beta.1](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/v1.0.26-beta.0...v1.0.26-beta.1) (2020-12-22)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-editor-core
+
 <a name="1.0.26-beta.0"></a>
 ## [1.0.26-beta.0](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/v1.0.25-beta.1...v1.0.26-beta.0) (2020-12-22)